+++ cfdm/cfdmimplementation.py
@@ -420,7 +420,7 @@ class CFDMImplementation(Implementation):
     def get_cell_connectivities(self, parent):
         """Return the cell connectivities from a parent.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters:
 
@@ -1032,7 +1032,7 @@ class CFDMImplementation(Implementation):
     def get_domain_topologies(self, parent):
         """Return the domain topologies from a parent.
 
-        .. versionadded:: (cfdm) NEXTVERSION
+        .. versionadded:: (cfdm) 1.13.1.0
 
         :Parameters:
 
@@ -2027,7 +2027,7 @@ class CFDMImplementation(Implementation):
             kwargs: optional
                 Parameters with which to initialise the object.
 
-                .. versionadded:: (cfdm) NEXTVERSION
+                .. versionadded:: (cfdm) 1.13.1.0
 
         :Returns: