+++ cfdm/data/pyfivearray.py
@@ -8,7 +8,7 @@ class PyfiveArray(IndexMixin, FileArray):
     Deprecated at version NEXTVERSION and is no longer available. Use
     `{{package}}.XnetcdfArray` instead.
 
-    .. versionadded:: (cfdm) NEXTVERSION
+    .. versionadded:: (cfdm) 1.13.1.0
 
     """