+++ cfdm/read_write/netcdf/cf_resolve_references.py
@@ -133,7 +133,10 @@ def resolve_reference(
         # Return the resoved reference
         return resolved_ref
 
-    # Reference couldn't be resolved, so return it unchanged.
+    # Reference couldn't be resolved, so return it unchanged. This
+    # could be the case for an external variable (e.g. the 'areacello'
+    # in cell_measures attribute 'area: areacello'; the 'sea' in the
+    # cell_methods attribute 'maximum where sea').
     return ref
 
 
@@ -369,6 +372,38 @@ def resolve_pattern_2(value, variable, coord=False):
 
 
 def resolve_pattern_3(value, variable, coord=False):
+    """Resolve references in a pattern 3 attribute.
+
+    Resolve references in an attribute whose value has one of the
+    following patterns:
+
+    * ''
+    * 'key1: var1'
+    * 'key1: var1 key2: var2'
+    * 'key1: var1 var2'
+    * 'key1: var1 var2 key2: var3'
+
+    E.g. ``cell_measures``, ``aggregated_data``, ``formula_terms``,
+    ``interpolation_parameters``
+
+    .. versionadded:: (cfdm) NEXTVERSION
+
+    """
+    try:
+        resolved = []
+        for ref in value.split():
+            if not ref.endswith(":"):
+                ref = resolve_reference(ref, variable, var=True, coord=coord)
+
+            resolved.append(ref)
+
+    except AttributeError:
+        return value
+    else:
+        return " ".join(resolved)
+
+
+def resolve_pattern_3b(value, variable, coord=False):
     """Resolve references in a pattern 3 attribute.
 
     Resolve references in an attribute whose value has one of the